added timer feature to the homepage with era-select screen
